move Builder to new Folder
